Added PlayerHealth script
Keeps track of playerHealth. Use DamagePlayer() to damage player by damageAmount. Use HealPlayer() to heal player by healAmount. It might be better to have one function that would take negative values for damage and positive values for healing. playerHealth is constrained between 0 and and maxPlayerHealth, but you can check whether the player was killed with IsPlayeralive(). We might also want to make the player health an integer instead of a float, depending on how we want damage to work.