Added CameraMovement script and prefab
Created CameraMovement script with three modes: stationary, centered, and mario. The mario mode has some play in it, so that the camera will only move if the player gets a certain distance from the center. I also made a camera prefab.
I also changed the damage and heal functions for player health so that they are publicly accessible.