Give player immunity and attach PlayerHealth script correctly
Updated PlayerHealth script so that DamagePlayer() can make the player immune for a few seconds. The player is also immune at the start of the game, which we can change. I changed the EnemyMovement script so that instead of having its own PlayerHealth, it instead gets the player's PlayerHealth component upon collision.