Added HealthPack and HealthPackSpawner prefabs
HealthPackSpawner spawns a new HealthPack at a set rate until a max limit on the amount of HealthPacks is reached. When the player collides with a HealthPack, they are healed, and the HealthPack is destroyed. Right now, the player physically collides with HealthPacks. If we want to change that, we can add a child to HealthPack with a Trigger collider on it, then put the script on the child. Then, we would set the parent to a layer that doesn't collide with the player. I have more details about this in a comment on the HealthPack script.